# Move type check of MD array to managed