… src/coreclr/vm/metasig.h