Removes a HELPER_METHOD_FRAME.

Generally there are more of MD array can be moved to managed, but not completely now.

| Method        | Job        | Toolchain         | Mean     | Error     | StdDev    | Ratio | RatioSD |
|-------------- |----------- |------------------ |---------:|----------:|----------:|------:|--------:|
| SetCovariant  | Job-KZUIWY | \PR\corerun.exe   | 2.758 ns | 0.0252 ns | 0.0236 ns |  0.99 |    0.01 |
| SetCovariant  | Job-ZSCMRL | \main\corerun.exe | 2.780 ns | 0.0191 ns | 0.0169 ns |  1.00 |    0.00 |
|               |            |                   |          |           |           |       |         |
| SetExactMatch | Job-KZUIWY | \PR\corerun.exe   | 2.768 ns | 0.0323 ns | 0.0302 ns |  0.99 |    0.02 |
| SetExactMatch | Job-ZSCMRL | \main\corerun.exe | 2.795 ns | 0.0246 ns | 0.0230 ns |  1.00 |    0.00 |

Not over optimizing, just ensuring no regression.