@jkotas 2024-06-16
> Not over optimizing, just ensuring no regression.

Could you please share the source for the benchmark? It is a bit suspicious that the difference between the baseline and the PR is in the noise range. I would expect it to be a small measurable improvement.