@huoyaoyuan 2024-06-16
Updated test:

```csharp
        private Base[,] baseArray = new Base[10, 1];
        private Base[,] covariantArray = new Derived[10, 1];
        private Derived[,] derivedArray = new Derived[10, 1];

        private Derived derivedObj = new Derived();

        private void GenericSet<T>(T[,] array, T value)
        {
            for (int i = 0; i < 10; i++)
            {
                array[i, 0] = value;
            }
        }

        [Benchmark]
        public void ReferenceTypeBase() => GenericSet(baseArray, derivedObj);

        [Benchmark]
        public void ReferenceTypeExactMatch() => GenericSet(derivedArray, derivedObj);

        [Benchmark]
        public void ReferenceTypeCovariant() => GenericSet(covariantArray, derivedObj);
```

Result:
| Method                  | Job        | Toolchain         | Mean     | Error    | StdDev   | Ratio | RatioSD |
|------------------------ |----------- |------------------ |---------:|---------:|---------:|------:|--------:|
| ReferenceTypeBase       | Job-IECQRX | \PR\corerun.exe   | 32.91 ns | 0.183 ns | 0.171 ns |  0.76 |    0.01 |
| ReferenceTypeBase       | Job-OMBWNE | \main\corerun.exe | 43.08 ns | 0.219 ns | 0.194 ns |  1.00 |    0.00 |
|                         |            |                   |          |          |          |       |         |
| ReferenceTypeExactMatch | Job-IECQRX | \PR\corerun.exe   | 29.27 ns | 0.422 ns | 0.395 ns |  1.00 |    0.02 |
| ReferenceTypeExactMatch | Job-OMBWNE | \main\corerun.exe | 29.13 ns | 0.342 ns | 0.320 ns |  1.00 |    0.00 |
|                         |            |                   |          |          |          |       |         |
| ReferenceTypeCovariant  | Job-IECQRX | \PR\corerun.exe   | 28.59 ns | 0.229 ns | 0.214 ns |  1.00 |    0.01 |
| ReferenceTypeCovariant  | Job-OMBWNE | \main\corerun.exe | 28.68 ns | 0.181 ns | 0.169 ns |  1.00 |    0.00 |

Yes it is measuable improvement when `array.GetType().ElementType` isn't exact `obj.GetType()`. The static type of array doesn't matter here.